+++ index.mjs
@@ -1,0 +1,2 @@
+import Client from "./client.mjs";
+module.exports = Client;
